infra/pyproject.toml
[project]
name = "infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-cdk-lib>=2.230.0",
    "constructs>=10.4.3",
]
